chore: refactor customization npcs
